requirements.txt
aiohttp==3.8.5
pandas==2.0.3
request==2.32.3
